Assumed Ratings Tweaks
Plex's genre capitalization didn't match for the 'kodomo' and '18 restricted' tags causing the content rating to not be set for series with said tags. Replaced the 'Mature' tag with 'borderline porn' as it doesn't seem to exist on anidb. 

Additionally, propose changing the 'TV-R' rating for '18 restricted' to 'X' since as far as I can tell that tag seems to be 100% porn on anidb and and an R rating would imply that parental supervision would allow minors to watch it.