Account for v3 API Series Search Endpoint Changes
v3 API Search results  are now more consistent and work with synonyms so account for this by using the order of the results to impact the score in Plex. This is also required since the distance will always be 0 for API searches that contain the same words as the search.