Append Additional Searchable Titles to the Sort Title
Append a user configurable 'Original Title' to the sort title to allow for another language to be searchable in Plex. This is quite helpful for series that are more commonly known by their english names than their romaji ones and vice versa.

Ideally the Original Title field would be used instead but due to a Plex issue it will not update in the UI when set through a custom agent.